I found that It will panic when I serialize and deserialize including custom type. I add some support for these types:
- uint8
- uint16
- uint32
- uint64
- int8
- int16
- int32
- int64