@ouromoros 2021-08-06
It looks correct to me now. Great!

 Last thing, I wonder if first calling `v.Int()` and then converting to the according type would have any performance impact. Perhaps should run a benchmark on it and see the difference.