@yihau 2021-08-06
here is the benchstat between master and my branch
```
name               old time/op  new time/op  delta
Deserialize-8      1.41µs ± 0%  1.69µs ± 0%   ~     (p=1.000 n=1+1)
Serialize-8         648ns ± 0%   623ns ± 0%   ~     (p=1.000 n=1+1)
FuzzSerialize-8    50.2µs ± 0%  46.6µs ± 0%   ~     (p=1.000 n=1+1)
FuzzDeserialize-8  4.65µs ± 0%  4.66µs ± 0%   ~     (p=1.000 n=1+1)
```

serialize is more efficient because get the value directly
there is a little bit performance impact to the deserialize function because convert